better application structure
