--- src/backend/backendServices/sendEmail.ts
@@ -1,34 +1,0 @@
-import { SendEmailCommand } from '@aws-sdk/client-ses';
-import { IRequestForAttestation } from '@kiltprotocol/types';
-
-import { sesClient } from './sesClient';
-
-export async function sendEmail(
-  url: string,
-  requestForAttestation: IRequestForAttestation,
-): Promise<void> {
-  const { contents } = requestForAttestation.claim;
-
-  const email = contents['Email'] as string;
-  const name = contents['Full name'] as string;
-
-  const params = {
-    Destination: {
-      ToAddresses: [email],
-    },
-    Source: 'test@socialkyc.io',
-    Message: {
-      Subject: {
-        Charset: 'UTF-8',
-        Data: 'SocialKYC - Confirm your email address',
-      },
-      Body: {
-        Text: {
-          Charset: 'UTF-8',
-          Data: `Hello ${name},\n\nThis is a test. Please click the link to confirm your email: ${url} \n\nKind regards,\nSocialKYC`,
-        },
-      },
-    },
-  };
-  await sesClient.send(new SendEmailCommand(params));
-}
