+++ src/backend/utilities/sesClient.ts
@@ -1,0 +1,8 @@
+import { SESClient } from '@aws-sdk/client-ses';
+
+import { configuration } from './configuration';
+
+export const sesClient = new SESClient({
+  region: configuration.aws.region,
+  credentials: configuration.aws,
+});
